agents/requirements.txt
# LiveKit Agent Framework
livekit-agents>=1.0.0
livekit-api>=0.5.0
livekit-plugins-assemblyai>=1.2.0
livekit-plugins-openai>=1.0.0
livekit-plugins-elevenlabs>=1.0.0

# Core dependencies
asyncio-mqtt>=0.16.1
python-dotenv>=1.0.0
pydantic>=2.5.0
websockets>=12.0
aiohttp>=3.10.0

# AI/ML libraries
openai>=1.6.0
anthropic>=0.8.0

# Audio processing
numpy>=1.24.3
scipy>=1.11.4

# Development tools
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
pytest>=7.4.3
pytest-asyncio>=0.23.2